requirements.txt
anthropic>=0.34.0
langgraph>=0.2.0
langchain-core>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
mcp>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.38.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0